# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для проекта Imperial",
        "title": "Imperial API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@imperial.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/v1/auth/confirm-email": {
            "post": {
                "description": "Подтверждение кода",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Подтверждение кода верификации на email",
                "parameters": [
                    {
                        "description": "Email для отправки кода верификации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ConfirmEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email успешно отправлен",
                        "schema": {
                            "$ref": "#/definitions/auth.ConfirmEmailResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/refresh-token": {
            "post": {
                "description": "Обновляет access токен с помощью refresh токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Обновление access токена",
                "parameters": [
                    {
                        "description": "Данные для обновления токена",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Токен успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/sign-in": {
            "post": {
                "description": "Аутентификация пользователя с помощью email и пароля",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Вход пользователя",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно аутентифицирован",
                        "schema": {
                            "$ref": "#/definitions/auth.SignInResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/sign-up": {
            "post": {
                "description": "Создает нового пользователя с указанным email и паролем",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегистрирован",
                        "schema": {
                            "$ref": "#/definitions/auth.SignUpResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/verify-email": {
            "post": {
                "description": "Отправляет код верификации на указанный email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Отправка кода верификации на email",
                "parameters": [
                    {
                        "description": "Email для отправки кода верификации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.VerifyEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email успешно отправлен",
                        "schema": {
                            "$ref": "#/definitions/auth.VerifyEmailResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.ConfirmEmailRequest": {
            "type": "object",
            "required": [
                "code",
                "email"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "123456"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
        "auth.ConfirmEmailResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email успешно подтвержден"
                }
            }
        },
        "auth.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "\u003crefresh_token\u003e"
                }
            }
        },
        "auth.RefreshResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "\u003caccess_token\u003e"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "\u003crefresh_token\u003e"
                }
            }
        },
        "auth.SignInRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "auth.SignInResponse": {
            "type": "object",
            "properties": {
                "tokens": {
                    "$ref": "#/definitions/entities.Tokens"
                },
                "user": {
                    "$ref": "#/definitions/auth.SignUpResponse"
                }
            }
        },
        "auth.SignUpRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "password123"
                }
            }
        },
        "auth.SignUpResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 123
                },
                "is_verified": {
                    "type": "boolean",
                    "example": false
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "auth.VerifyEmailRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                }
            }
        },
        "auth.VerifyEmailResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Email успешно отправлен"
                }
            }
        },
...
        "entities.Tokens": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token must be passed with `Bearer ` prefix. Example: \"Bearer eyJhbGciOiJIUzI1NiI...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}